--- src/main/java/com/codeit/duckhu/comments/conntroller/CommentController.java
@@ -1,12 +1,0 @@
-package com.codeit.duckhu.comments.conntroller;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/api/comments")
-@RequiredArgsConstructor
-public class CommentController {
-
-}
