--- src/main/java/com/codeit/duckhu/comments/dto/request/CommentCreateRequest.java
@@ -1,11 +1,0 @@
-package com.codeit.duckhu.comments.dto.request;
-
-import java.util.UUID;
-import lombok.Data;
-
-@Data
-public class CommentCreateRequest {
-  UUID userId;
-  UUID reviewId;
-  String content;
-}
