+++ src/main/java/com/codeit/duckhu/domain/comments/conntroller/CommentController.java
@@ -1,0 +1,12 @@
+package com.codeit.duckhu.domain.comments.conntroller;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/api/comments")
+@RequiredArgsConstructor
+public class CommentController {
+
+}
