+++ src/main/java/com/codeit/duckhu/domain/comments/dto/request/CommentCreateRequest.java
@@ -1,0 +1,11 @@
+package com.codeit.duckhu.domain.comments.dto.request;
+
+import java.util.UUID;
+import lombok.Data;
+
+@Data
+public class CommentCreateRequest {
+  UUID userId;
+  UUID reviewId;
+  String content;
+}
