--- src/main/java/com/codeit/duckhu/review/exception/ReviewErrorCode.java
@@ -1,23 +1,0 @@
-package com.codeit.duckhu.review.exception;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
-
-/**
- * Custom Error 지정 필요성이 있다면 에러 코드를 작성하여 진행.
- * <p>
- * 일반적인 상황의 에러만 적혀있지만, 특정 도메인에서 발생할 수 있는 에러 등 정의\n\n
- * CustomException(ErrorCode.INTERNAL_SERVER_ERROR) 처럼 사용.
- */
-@Getter
-@RequiredArgsConstructor
-public enum ReviewErrorCode {
-  // 리뷰 에러
-  REVIEW_NOT_FOUND(HttpStatus.NOT_FOUND, "리뷰를 찾을 수 없습니다.", "리뷰를 찾을 수 없습니다.");
-
-
-  private final HttpStatus status;
-  private final String message;
-  private final String detail;
-}
