--- src/test/java/com/codeit/duckhu/comments/CommentTest.java
@@ -1,19 +1,0 @@
-package com.codeit.duckhu.comments;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
-import com.codeit.duckhu.comments.domain.Comment;
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-public class CommentTest {
-
-  @Test
-  void get(){
-        User user = new User();
-        Comment comment = new Comment(user,new Review(),"content test");
-
-        assertEquals(user, comment.getUser());
-  }
-}
