+++ src/test/java/com/codeit/duckhu/domain/comments/CommentTest.java
@@ -1,0 +1,19 @@
+package com.codeit.duckhu.domain.comments;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+import com.codeit.duckhu.comments.domain.Comment;
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+public class CommentTest {
+
+  @Test
+  void get(){
+        User user = new User();
+        Comment comment = new Comment(user,new Review(),"content test");
+
+        assertEquals(user, comment.getUser());
+  }
+}
